+++ test/extensions/persistent_session/phoenix/controllers/controller_callbacks_test.exs
@@ -10,7 +10,7 @@ defmodule PowPersistentSession.Phoenix.ControllerCallbacksTest do
   describe "Pow.Phoenix.SessionController.create/2" do
     test "generates cookie", %{conn: conn, ets: ets} do
       conn = post conn, Routes.pow_session_path(conn, :create, %{"user" => @valid_params})
-      assert session_fingerprint = conn.private[:pow_session_fingerprint]
+      assert session_fingerprint = conn.private[:pow_session_metadata][:fingerprint]
 
       assert %{max_age: @max_age, path: "/", value: id} = conn.resp_cookies[@cookie_key]
       assert PersistentSessionCache.get([backend: ets], id) == {1, session_fingerprint: session_fingerprint}
@@ -29,7 +29,6 @@ defmodule PowPersistentSession.Phoenix.ControllerCallbacksTest do
 
       assert conn.resp_cookies[@cookie_key]
     end
-
   end
 
   describe "Pow.Phoenix.SessionController.delete/2" do