+++ test/pow/plug/session_test.exs
@@ -55,7 +55,24 @@ defmodule Pow.Plug.SessionTest do
       |> Session.call(opts)
 
     assert conn.assigns[:current_user] == @user
-    assert conn.private[:pow_session_fingerprint] == "fingerprint"
+    assert conn.private[:pow_session_metadata][:fingerprint] == "fingerprint"
+  end
+
+  test "call/2 with stored session and custom metadata", %{conn: conn} do
+    inserted_at = :os.system_time(:millisecond)
+    CredentialsCache.put(@store_config, "token", {@user, inserted_at: inserted_at, a: 1})
+
+    opts = Session.init(@default_opts)
+    conn =
+      conn
+      |> Conn.put_private(:pow_session_metadata, b: 2)
+      |> Conn.fetch_session()
+      |> Conn.put_session(@default_opts[:session_key], "token")
+      |> Session.call(opts)
+
+    assert conn.assigns[:current_user] == @user
+    assert conn.private[:pow_session_metadata][:inserted_at] == inserted_at
+    assert conn.private[:pow_session_metadata][:a] == 1
   end
 
   test "call/2 with non existing cached key", %{conn: conn} do
@@ -100,7 +117,7 @@ defmodule Pow.Plug.SessionTest do
     assert {_user, metadata} = CredentialsCache.get(@store_config, new_session_id)
     assert metadata[:inserted_at] != stale_timestamp
     assert metadata[:fingerprint] == "fingerprint"
-    assert conn.private[:pow_session_fingerprint] == "fingerprint"
+    assert conn.private[:pow_session_metadata][:fingerprint] == "fingerprint"
   end
 
   test "call/2 with prepended `:otp_app` session key", %{conn: conn} do
@@ -168,6 +185,21 @@ defmodule Pow.Plug.SessionTest do
       assert CredentialsCache.get(@store_config, session_id) == :not_found
       assert Plug.current_user(conn) == @user
       assert metadata[:fingerprint] == new_metadata[:fingerprint]
+    end
+
+    test "creates with custom metadata", %{conn: conn} do
+      inserted_at = :os.system_time(:millisecond) - 10
+      opts = Session.init(@default_opts)
+      conn =
+        conn
+        |> Conn.put_private(:pow_session_metadata, inserted_at: inserted_at, a: 1)
+        |> Session.call(opts)
+        |> Session.do_create(@user, opts)
+
+      assert conn.assigns[:current_user] == @user
+      assert conn.private[:pow_session_metadata][:inserted_at] != inserted_at
+      assert conn.private[:pow_session_metadata][:fingerprint]
+      assert conn.private[:pow_session_metadata][:a] == 1
     end
 
     test "creates new session id with `:otp_app` prepended", %{conn: conn} do