+++ CHANGELOG.md
@@ -39,8 +39,8 @@
   * `Pow.Store.CredentialsCache` now adds a session key rather than appending to a list for the user key to prevent race condition
 * Fixed bug where `PowEmailConfirmation.Phoenix.ControllerCallbacks` couldn't deliver email
 * `Pow.Plug.Session` now adds a fingerprint for new sessions that will persist when sessions are renewed
-* `Pow.Plug.Session` assigns private key `:pow_session_fingerprint` in the conn with the session fingerprint value, or uses the private key value when creating sessions
-* `PowPersistentSession.Plug.Cookie` will record the value of `:pow_session_fingerprint` if it exists in conn private, and will assign it to the conn if `:session_fingerprint` exists for the persistent session metadata
+* `Pow.Plug.Session` assigns private key `:pow_session_metadata` in the conn as a keyword list with `:fingerprint` containing the session fingerprint value when fetching session, and fetches this value when creating sessions
+* `PowPersistentSession.Plug.Cookie` will record the value of `:fingerprint` in the `:pow_session_metadata` if it exists in conn private, and will assign it to the conn if `:session_fingerprint` exists for the persistent session metadata
 
 ## v1.0.13 (2019-08-25)